Refactor PropertyChangedEventArgsUnnecessaryAllocation
