+++ test/CSharp/CodeCracker.Test/CodeCracker.Test.csproj
@@ -130,8 +130,7 @@
     <Compile Include="Refactoring\IntroduceFieldFromConstructorTest.cs" />
     <Compile Include="Refactoring\ComputeExpressionTests.cs" />
     <Compile Include="Refactoring\ReplaceWithGetterOnlyAutoPropertyTests.cs" />
-    <Compile Include="Refactoring\PropertyChangedEventArgsUnnecessaryAllocationAnalyzerTests.cs" />
-    <Compile Include="Refactoring\PropertyChangedEventArgsUnnecessaryAllocationCodeFixProviderTests.cs" />
+    <Compile Include="Refactoring\PropertyChangedEventArgsUnnecessaryAllocationTests.cs" />
     <Compile Include="Refactoring\SplitIntoNestedIfTests.cs" />
     <Compile Include="Refactoring\ParameterRefectoryTests.cs" />
     <Compile Include="Refactoring\AllowMembersOrderingCodeFixProviderTests.StyleCop.cs" />
